Refactored WikiContentProvider into ModifyingPageRetriever
Like this it's a decorator, and is never bound to expect in the factory.

First step towards https://phabricator.wikimedia.org/T138964
